+++ include/boundaryInit.h
@@ -15,34 +15,34 @@ using namespace std;
 #include <cmath>
 
 struct boundary_init {
-    float background_Z;
-    float background_amu;
+    double background_Z;
+    double background_amu;
     int nR_Temp;
     int nZ_Temp;
-    float* TempGridr;
-    float* TempGridz;
-    float* ti;
-    float* te;
+    double* TempGridr;
+    double* TempGridz;
+    double* ti;
+    double* te;
     int nx;
     int nz;
-    float* densityGridx;
-    float* densityGridz;
-    float* density;
-    float* ne;
+    double* densityGridx;
+    double* densityGridz;
+    double* density;
+    double* ne;
     int nxB;
     int nzB;
-    float* bfieldGridr;
-    float* bfieldGridz;
-    float* bfieldR;
-    float* bfieldZ;
-    float* bfieldT;
-    float potential;
+    double* bfieldGridr;
+    double* bfieldGridz;
+    double* bfieldR;
+    double* bfieldZ;
+    double* bfieldT;
+    double potential;
     
-    boundary_init(float _background_Z, float _background_amu,int _nx, int _nz,
-          float* _densityGridx, float* _densityGridz,float* _density,float* _ne,int _nxB,
-          int _nzB, float* _bfieldGridr, float* _bfieldGridz,float* _bfieldR,
-          float* _bfieldZ,float* _bfieldT,int _nR_Temp, int _nZ_Temp,
-          float* _TempGridr, float* _TempGridz, float* _ti, float* _te, float _potential)
+    boundary_init(double _background_Z, double _background_amu,int _nx, int _nz,
+          double* _densityGridx, double* _densityGridz,double* _density,double* _ne,int _nxB,
+          int _nzB, double* _bfieldGridr, double* _bfieldGridz,double* _bfieldR,
+          double* _bfieldZ,double* _bfieldT,int _nR_Temp, int _nZ_Temp,
+          double* _TempGridr, double* _TempGridz, double* _ti, double* _te, double _potential)
 
      : background_Z(_background_Z),
         background_amu(_background_amu),
@@ -69,39 +69,39 @@ struct boundary_init {
 
     void operator()(Boundary &b) const {
 #if USE3DTETGEOM
-        float midpointx = b.x1 + 0.666666667*(b.x2 + 0.5*(b.x3-b.x2)-b.x1);
-        float midpointy = b.y1 + 0.666666667*(b.y2 + 0.5*(b.y3-b.y2)-b.y1);
-        float midpointz = b.z1 + 0.666666667*(b.z2 + 0.5*(b.z3-b.z2)-b.z1);
+        double midpointx = b.x1 + 0.666666667*(b.x2 + 0.5*(b.x3-b.x2)-b.x1);
+        double midpointy = b.y1 + 0.666666667*(b.y2 + 0.5*(b.y3-b.y2)-b.y1);
+        double midpointz = b.z1 + 0.666666667*(b.z2 + 0.5*(b.z3-b.z2)-b.z1);
         b.midx = midpointx;
         b.midy = midpointy;
         b.midz = midpointz;
 #else
 
-        float midpointx = 0.5*(b.x2 - b.x1)+ b.x1;
-        float midpointy = 0.0;
-        float midpointz = 0.5*(b.z2 - b.z1) + b.z1;
+        double midpointx = 0.5*(b.x2 - b.x1)+ b.x1;
+        double midpointy = 0.0;
+        double midpointz = 0.5*(b.z2 - b.z1) + b.z1;
 #endif
         b.density = interp2dCombined(midpointx,midpointy,midpointz,nx,nz,densityGridx,densityGridz,density);
         b.ne = interp2dCombined(midpointx,midpointy,midpointz,nx,nz,densityGridx,densityGridz,ne);
         b.ti = interp2dCombined(midpointx,midpointy,midpointz,nR_Temp,nZ_Temp,TempGridr,TempGridz,ti);
         b.te = interp2dCombined(midpointx,midpointy,midpointz,nR_Temp,nZ_Temp,TempGridr,TempGridz,te);
-        float B[3] = {0.0,0.0,0.0};
+        double B[3] = {0.0,0.0,0.0};
 interp2dVector(&B[0],midpointx,midpointy,midpointz,nxB,nzB,bfieldGridr,
                  bfieldGridz,bfieldR,bfieldZ,bfieldT);
-        float norm_B = vectorNorm(B);
+        double norm_B = vectorNorm(B);
 #if USE3DTETGEOM
-        float surfNorm[3] = {0.0,0.0,0.0};
+        double surfNorm[3] = {0.0,0.0,0.0};
         b.getSurfaceNormal(surfNorm,0.0,0.0);
-        float theta = acos(vectorDotProduct(B,surfNorm)/(vectorNorm(B)*vectorNorm(surfNorm)));
+        double theta = acos(vectorDotProduct(B,surfNorm)/(vectorNorm(B)*vectorNorm(surfNorm)));
         if (theta > 3.14159265359*0.5)
         {
           theta = abs(theta - (3.14159265359));
         }
 #else
-        float br = B[0];
-        float bt = B[1];
-        float bz = B[2];
-        float theta = acos((-br*b.slope_dzdx + bz)/(sqrt(br*br+bz*bz+bt*bt)*sqrt(b.slope_dzdx*b.slope_dzdx + 1.0)));
+        double br = B[0];
+        double bt = B[1];
+        double bz = B[2];
+        double theta = acos((-br*b.slope_dzdx + bz)/(sqrt(br*br+bz*bz+bt*bt)*sqrt(b.slope_dzdx*b.slope_dzdx + 1.0)));
  
         if (theta > 3.14159265359*0.5)
         {
@@ -116,8 +116,8 @@ interp2dVector(&B[0],midpointx,midpointy,midpointz,nxB,nzB,bfieldGridr,
         b.impacts = 0.0;
 #if BIASED_SURFACE
         b.potential = potential;
-        //float cs = sqrt(2*b.ti*1.602e-19/(1.66e-27*background_amu));
-        //float jsat_ion = 1.602e-19*b.density*cs;
+        //double cs = sqrt(2*b.ti*1.602e-19/(1.66e-27*background_amu));
+        //double jsat_ion = 1.602e-19*b.density*cs;
         //b.ChildLangmuirDist = 2.0/3.0*pow(2*1.602e-19/(background_amu*1.66e-27),0.25)
         //*pow(potential,0.75)/(2.0*sqrt(3.1415*jsat_ion))*1.055e-5;
         if(b.te > 0.0)
